crud denounce

+++ src/server/src/KNTC.Application.Contracts/FileAttachments/FileAttachmentDto.cs
@@ -8,7 +8,7 @@ public class FileAttachmentDto : AuditedEntityDto<Guid>
     public Guid? ComplainId { get; set; }
     public Guid? DenounceId { get; set; }
     public string TenTaiLieu { get; private set; }
-    public short GiaiDoan { get; set; }
+    public int GiaiDoan { get; set; }
     public int HinhThuc { get; set; }
     public DateTime ThoiGianBanHanh { get; set; }
     public DateTime NgayNhan { get; set; }
